Added simplified CT-to-CT driver script.
This script permits the inputs and outputs to be more easily specified compared to the currently used script
which has rigid input naming conventions and filesystem layout. The earlier script also assumed multiple volumes
were being registered to a single fixed volume. In contrast, this script is meant to be invoked once per image
volume, and inputs/outputs can be located anywhere.