Added a bspline-based default file.
The Demons algorithm was causing issues on some files, complaining about non-overlapping images even when images
appeared to significantly spatially overlap (e.g., immediately after a centre alignment). Bsplines do not appear
to suffer the same issue. Note that the Demons algorithm is implemented in ITK whereas the BSplines is in Plastimatch.