De-identifier: support optional arguments.
This commits adds support for optional parameters; currently only study and series descriptions.

If interactive queries time-out, then the default value will be accepted. If the parameter is optional, it will not be
set. At the moment, waiting for the timeout (or possible [ctrl]+[d]) is required to avoid specifiying an optional
parameter.